+++ src/edyn/shapes/triangle_shape.cpp
@@ -35,11 +35,11 @@ triangle_edges get_triangle_edges(const triangle_vertices &vertices) {
 void get_triangle_support_feature(const triangle_vertices &vertices, 
                                   const vector3 &axis_pos, const vector3 &axis_dir,
                                   triangle_feature &tri_feature,
-                                  uint8_t &tri_feature_index,
+                                  size_t &tri_feature_index,
                                   scalar &projection) {
     projection = -large_scalar;
 
-    for (uint8_t i = 0; i < 3; ++i) {
+    for (size_t i = 0; i < 3; ++i) {
         auto &v = vertices[i];
         auto proj_i = dot(v - axis_pos, axis_dir);
